// src/app/generated/dcc/model/statementDto.ts
/**
 * 
 * 
 *
 * 
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { LanguageSpecificStringsDto } from './languageSpecificStringsDto';
import { ContactDto } from './contactDto';
import { DataDto } from './dataDto';
import { RichContentDto } from './richContentDto';
import { LocationDto } from './locationDto';


export interface StatementDto { 
    id?: string;
    refIds?: Array<string>;
    refTypes?: Array<string>;
    data?: Array<DataDto>;
    countryCodes?: Array<string>;
    convention?: string;
    traceable?: boolean;
    conformity?: string;
    conformityXMLList?: Array<string>;
    date?: string;
    period?: string;
    name?: LanguageSpecificStringsDto;
    norms?: Array<string>;
    references?: Array<string>;
    description?: RichContentDto;
    declaration?: RichContentDto;
    valid?: boolean;
    validXMLList?: Array<boolean>;
    nonSIDefinition?: string;
    nonSIUnit?: string;
    location?: LocationDto;
    responsibleAuthority?: ContactDto;
}