// src/app/generated/dcc/model/languageSpecificStringsDto.ts
/**
 * 
 * 
 *
 * 
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { LangTextPair } from './langTextPair';


export interface LanguageSpecificStringsDto { 
    id?: string;
    refIds?: Array<string>;
    content?: Array<LangTextPair>;
}